split unittest and integration test on reporter api
